refactor(hsm): split more into modules
docs(hsm): cleanup some docs
