+++ zcash-hsmbuilder/Cargo.toml
@@ -35,7 +35,6 @@ secp256k1 = { version = "0.20" }
 sha2 = "0.9.1"
 ripemd160 = "0.9.1"
 
-serde_derive = "1.0.117"
 serde = { version = "1.0.117", features = ["derive"] }
 
 #zcash
