+++ zcash-hsmbuilder/src/data/neon_bridge.rs
