+++ zcash-hsmbuilder/src/txbuilder.rs
@@ -1,34 +1,38 @@
 //! Structs for building transactions.
-use std::io::{self, Write};
-use std::marker::PhantomData;
+use std::{
+    io::{self, Write},
+    marker::PhantomData,
+};
 
+use crypto_api_chachapoly::ChachaPolyIetf;
 use group::GroupEncoding;
 use rand::{rngs::OsRng, CryptoRng, RngCore};
-use zcash_primitives::constants::SPENDING_KEY_GENERATOR;
-use zcash_primitives::note_encryption::SaplingNoteEncryption;
-use zcash_primitives::primitives::{Diversifier, Note, PaymentAddress, ProofGenerationKey, Rseed};
-use zcash_primitives::redjubjub::Signature;
-use zcash_primitives::transaction::components::amount::DEFAULT_FEE;
-use zcash_primitives::transaction::components::{Amount, TxOut, GROTH_PROOF_SIZE};
-use zcash_primitives::transaction::{
-    signature_hash_data, SignableInput, Transaction, TransactionData, SIGHASH_ALL,
-};
 use zcash_primitives::{
-    consensus, keys::OutgoingViewingKey, legacy::TransparentAddress, memo::MemoBytes as Memo,
-    merkle_tree::MerklePath, redjubjub::PublicKey, sapling::Node, util::generate_random_rseed,
-};
-use zcash_primitives::{
-    legacy::Script,
-    transaction::components::{OutPoint, TxIn},
+    consensus,
+    constants::SPENDING_KEY_GENERATOR,
+    keys::OutgoingViewingKey,
+    legacy::{Script, TransparentAddress},
+    memo::MemoBytes as Memo,
+    merkle_tree::MerklePath,
+    note_encryption::SaplingNoteEncryption,
+    primitives::{Diversifier, Note, PaymentAddress, ProofGenerationKey, Rseed},
+    redjubjub::{PublicKey, Signature},
+    sapling::Node,
+    transaction::{
+        components::{amount::DEFAULT_FEE, Amount, OutPoint, TxIn, TxOut, GROTH_PROOF_SIZE},
+        signature_hash_data, SignableInput, Transaction, TransactionData, SIGHASH_ALL,
+    },
+    util::generate_random_rseed,
 };
 
-use sha2::{Digest, Sha256};
+use crate::{
+    data::{sighashdata::signature_hash_input_data, HashSeed, HsmTxData},
+    errors::Error,
+    txprover::HsmTxProver,
+};
 
-use crate::errors::Error;
-use crate::sighashdata::signature_hash_input_data;
-use crate::txprover::TxProver;
-use crate::{HashSeed, HsmTxData};
-use crypto_api_chachapoly::ChachaPolyIetf;
+mod builder_data;
+pub use builder_data::*;
 
 const DEFAULT_TX_EXPIRY_DELTA: u32 = 20;
 
@@ -36,268 +40,10 @@ const DEFAULT_TX_EXPIRY_DELTA: u32 = 20;
 /// with dummy outputs if necessary. See https://github.com/zcash/zcash/issues/3615
 const MIN_SHIELDED_OUTPUTS: usize = 2;
 
-const OUT_PLAINTEXT_SIZE: usize = 32 + // pk_d
-    32; // esk
-const OUT_CIPHERTEXT_SIZE: usize = OUT_PLAINTEXT_SIZE + 16;
-
-#[derive(Clone)]
-struct SpendDescriptionInfo {
-    //extsk: ExtendedSpendingKey, //change this to path in ledger
-    diversifier: Diversifier,
-    note: Note,
-    alpha: jubjub::Fr,
-    //get both from ledger and generate self
-    merkle_path: MerklePath<Node>,
-    proofkey: ProofGenerationKey,
-    //get from ledger
-    rcv: jubjub::Fr,
-}
-
-#[derive(Clone)]
-pub struct SaplingOutput {
-    /// `None` represents the `ovk = ⊥` case.
-    ovk: Option<OutgoingViewingKey>,
-    //get from ledger
-    to: PaymentAddress,
-    note: Note,
-    memo: Memo,
-    rcv: jubjub::Fr, //get from ledger
-    hashseed: Option<HashSeed>,
-}
-
-impl SaplingOutput {
-    pub fn new<R: RngCore + CryptoRng, P: consensus::Parameters>(
-        ovk: Option<OutgoingViewingKey>,
-        to: PaymentAddress,
-        value: Amount,
-        memo: Option<Memo>,
-        rcv: jubjub::Fr,
-        rseed: Rseed,
-        hashseed: Option<HashSeed>,
-    ) -> Result<Self, Error> {
-        let g_d = match to.g_d() {
-            Some(g_d) => g_d,
-            None => return Err(Error::InvalidAddress),
-        };
-        if value.is_negative() {
-            return Err(Error::InvalidAmount);
-        }
-
-        //let rseed = generate_random_rseed::<P, R>(height, rng);
-
-        let note = Note {
-            g_d,
-            pk_d: *to.pk_d(),
-            value: value.into(),
-            rseed,
-        };
-
-        Ok(SaplingOutput {
-            ovk,
-            to,
-            note,
-            memo: memo.unwrap_or_else(Memo::empty),
-            rcv,
-            hashseed,
-        })
-    }
-
-    pub fn build<P: TxProver, R: RngCore + CryptoRng>(
-        self,
-        prover: &P,
-        ctx: &mut P::SaplingProvingContext,
-        rng: &mut R,
-    ) -> zcash_primitives::transaction::components::OutputDescription {
-        let mut encryptor = SaplingNoteEncryption::new(
-            self.ovk,
-            self.note.clone(),
-            self.to.clone(),
-            self.memo,
-            rng,
-        );
-
-        let (zkproof, cv) = prover.output_proof(
-            ctx,
-            *encryptor.esk(),
-            self.to,
-            self.note.rcm(),
-            self.note.value,
-            self.rcv,
-        );
-
-        let cmu = self.note.cmu();
-
-        let enc_ciphertext = encryptor.encrypt_note_plaintext();
-        let mut out_ciphertext;
-        if self.ovk.is_some() {
-            out_ciphertext = encryptor.encrypt_outgoing_plaintext(&cv, &cmu);
-        } else {
-            let seed = self.hashseed.unwrap().0;
-            let mut randbytes = [0u8; 32 + OUT_PLAINTEXT_SIZE];
-            for i in 0..3 {
-                let mut sha256 = Sha256::new();
-                sha256.update([i as u8]);
-                sha256.update(seed);
-                let h = sha256.finalize();
-                randbytes[i * 32..(i + 1) * 32].copy_from_slice(&h);
-            }
-            let mut ock = [0u8; 32];
-            ock.copy_from_slice(&randbytes[0..32]);
-            let mut input = [0u8; OUT_PLAINTEXT_SIZE];
-            input.copy_from_slice(&randbytes[32..]);
-            out_ciphertext = [0u8; OUT_CIPHERTEXT_SIZE];
-            assert_eq!(
-                ChachaPolyIetf::aead_cipher()
-                    .seal_to(&mut out_ciphertext, &input, &[], ock.as_ref(), &[0u8; 12])
-                    .unwrap(),
-                OUT_CIPHERTEXT_SIZE
-            );
-        }
-
-        let ephemeral_key = (*encryptor.epk()).into();
-
-        zcash_primitives::transaction::components::OutputDescription {
-            cv,
-            cmu,
-            ephemeral_key,
-            enc_ciphertext,
-            out_ciphertext,
-            zkproof,
-        }
-    }
-}
-
-struct TransparentInputInfo {
-    pubkey: secp256k1::PublicKey,
-    coin: TxOut,
-}
-
-struct TransparentInputs {
-    secp: secp256k1::Secp256k1<secp256k1::VerifyOnly>,
-    inputs: Vec<TransparentInputInfo>,
-}
-
-impl Default for TransparentInputs {
-    fn default() -> Self {
-        TransparentInputs {
-            secp: secp256k1::Secp256k1::gen_new(),
-            inputs: Default::default(),
-        }
-    }
-}
-
-impl TransparentInputs {
-    fn push(
-        &mut self,
-        mtx: &mut TransactionData,
-        pubkey: secp256k1::PublicKey,
-        utxo: OutPoint,
-        coin: TxOut,
-    ) -> Result<(), Error> {
-        if coin.value.is_negative() {
-            return Err(Error::InvalidAmount);
-        }
-        match coin.script_pubkey.address() {
-            Some(TransparentAddress::PublicKey(hash)) => {
-                use ripemd160::Ripemd160;
-                use sha2::{Digest, Sha256};
-                if hash[..] != Ripemd160::digest(&Sha256::digest(&pubkey.serialize()))[..] {
-                    return Err(Error::InvalidAddressHash);
-                }
-            }
-            _ => return Err(Error::InvalidAddressFormat),
-        }
-
-        mtx.vin.push(TxIn::new(utxo));
-        self.inputs.push(TransparentInputInfo { pubkey, coin });
-
-        Ok(())
-    }
-
-    fn value_sum(&self) -> Amount {
-        {
-            self.inputs
-                .iter()
-                .map(|input| input.coin.value)
-                .sum::<Amount>()
-        }
-    }
-    fn apply_signatures(
-        &self,
-        signatures: Vec<secp256k1::Signature>,
-        mtx: &mut TransactionData,
-        consensus_branch_id: consensus::BranchId,
-    ) -> Result<(), Error> {
-        if signatures.len() != self.inputs.len() {
-            return Err(Error::TranspararentSig);
-        }
-        let mut sighash = [0u8; 32];
-
-        for (i, info) in self.inputs.iter().enumerate() {
-            sighash.copy_from_slice(&signature_hash_data(
-                mtx,
-                consensus_branch_id,
-                SIGHASH_ALL,
-                SignableInput::transparent(i, &info.coin.script_pubkey, info.coin.value),
-            ));
-
-            let msg = secp256k1::Message::from_slice(&sighash).expect("32 bytes");
-            let sig = signatures[i];
-            let pk = info.pubkey;
-            if self.secp.verify(&msg, &sig, &pk).is_err() {
-                return Err(Error::TranspararentSig);
-            }
-            // Signature has to have "SIGHASH_ALL" appended to it
-            let mut sig_bytes: Vec<u8> = sig.serialize_der()[..].to_vec();
-            sig_bytes.extend(&[SIGHASH_ALL as u8]);
-
-            // P2PKH scriptSig
-            mtx.vin[i].script_sig =
-                Script::default() << &sig_bytes[..] << &info.pubkey.serialize()[..];
-        }
-        Ok(())
-    }
-}
-
-/// Metadata about a transaction created by a [`Builder`].
-#[derive(Debug, PartialEq, Clone)]
-pub struct TransactionMetadata {
-    spend_indices: Vec<usize>,
-    output_indices: Vec<usize>,
-}
-
-impl TransactionMetadata {
-    fn new() -> Self {
-        TransactionMetadata {
-            spend_indices: vec![],
-            output_indices: vec![],
-        }
-    }
-
-    /// Returns the index within the transaction of the [`SpendDescription`] corresponding
-    /// to the `n`-th call to [`Builder::add_sapling_spend`].
-    ///
-    /// Note positions are randomized when building transactions for indistinguishability.
-    /// This means that the transaction consumer cannot assume that e.g. the first spend
-    /// they added (via the first call to [`Builder::add_sapling_spend`]) is the first
-    /// [`SpendDescription`] in the transaction.
-    pub fn spend_index(&self, n: usize) -> Option<usize> {
-        self.spend_indices.get(n).copied()
-    }
-
-    /// Returns the index within the transaction of the [`OutputDescription`] corresponding
-    /// to the `n`-th call to [`Builder::add_sapling_output`].
-    ///
-    /// Note positions are randomized when building transactions for indistinguishability.
-    /// This means that the transaction consumer cannot assume that e.g. the first output
-    /// they added (via the first call to [`Builder::add_sapling_output`]) is the first
-    /// [`OutputDescription`] in the transaction.
-    pub fn output_index(&self, n: usize) -> Option<usize> {
-        self.output_indices.get(n).copied()
-    }
-}
-
 /// Generates a [`Transaction`] from its inputs and outputs.
+///
+/// This is a rather low level builder, and is a HSM-compatible version
+/// of [`zcash_primitives::transaction::builder::Builder`].
 pub struct Builder<P: consensus::Parameters, R: RngCore + CryptoRng> {
     rng: R,
     height: u32,
@@ -312,7 +58,7 @@ pub struct Builder<P: consensus::Parameters, R: RngCore + CryptoRng> {
 }
 
 impl<P: consensus::Parameters> Builder<P, OsRng> {
-    /// Creates a new `Builder` targeted for inclusion in the block with the given height,
+    /// Creates a new [`Builder`] targeted for inclusion in the block with the given height,
     /// using default values for general transaction fields and the default OS random.
     ///
     /// # Default values
@@ -330,167 +76,8 @@ impl<P: consensus::Parameters> Builder<P, OsRng> {
     }
 }
 
-fn spend_old_data_fromtx(data: &[SpendDescriptionInfo]) -> Vec<NullifierInput> {
-    let mut v = Vec::new();
-    for info in data.iter() {
-        let n = NullifierInput {
-            rcm_old: info.note.rcm().to_bytes(),
-            notepos: info.merkle_path.position.to_le_bytes(),
-        };
-        v.push(n);
-    }
-    v
-}
-
-fn transparent_script_data_fromtx(
-    tx: &TransactionData,
-    inputs: &[TransparentInputInfo],
-) -> Result<Vec<TransparentScriptData>, Error> {
-    let mut data = Vec::new();
-    for (i, info) in inputs.iter().enumerate() {
-        let mut prevout = [0u8; 36];
-        prevout[0..32].copy_from_slice(tx.vin[i].prevout.hash().as_ref());
-        prevout[32..36].copy_from_slice(&tx.vin[i].prevout.n().to_le_bytes());
-
-        let mut script_pubkey = [0u8; 26];
-        info.coin.script_pubkey.write(&mut script_pubkey[..])?;
-
-        let mut value = [0u8; 8];
-        value.copy_from_slice(&info.coin.value.to_i64_le_bytes());
-
-        let mut sequence = [0u8; 4];
-        sequence.copy_from_slice(&tx.vin[i].sequence.to_le_bytes());
-
-        let ts = TransparentScriptData {
-            prevout,
-            script_pubkey,
-            value,
-            sequence,
-        };
-        data.push(ts);
-    }
-    Ok(data)
-}
-
-fn spend_data_hms_fromtx(
-    input: &[zcash_primitives::transaction::components::SpendDescription],
-) -> Vec<SpendDescription> {
-    let mut data = Vec::new();
-    for info in input.iter() {
-        let description = SpendDescription::from(info);
-        data.push(description);
-    }
-    data
-}
-
-fn output_data_hsm_fromtx(
-    input: &[zcash_primitives::transaction::components::OutputDescription],
-) -> Vec<OutputDescription> {
-    let mut data = Vec::new();
-    for info in input.iter() {
-        let description = OutputDescription::from(info);
-        data.push(description);
-    }
-    data
-}
-
-#[derive(Clone)]
-pub struct NullifierInput {
-    pub rcm_old: [u8; 32],
-    pub notepos: [u8; 8],
-}
-
-impl NullifierInput {
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_all(&self.rcm_old)?;
-        writer.write_all(&self.notepos)
-    }
-}
-
-#[derive(Clone)]
-pub struct TransparentScriptData {
-    pub prevout: [u8; 36],
-    pub script_pubkey: [u8; 26],
-    pub value: [u8; 8],
-    pub sequence: [u8; 4],
-}
-
-impl TransparentScriptData {
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_all(&self.prevout)?;
-        writer.write_all(&self.script_pubkey)?;
-        writer.write_all(&self.value)?;
-        writer.write_all(&self.sequence)
-    }
-}
-
-#[derive(Clone)]
-pub struct SpendDescription {
-    pub cv: [u8; 32],
-    pub anchor: [u8; 32],
-    pub nullifier: [u8; 32],
-    pub rk: [u8; 32],
-    pub zkproof: [u8; GROTH_PROOF_SIZE],
-}
-
-impl SpendDescription {
-    pub fn from(
-        info: &zcash_primitives::transaction::components::SpendDescription,
-    ) -> SpendDescription {
-        SpendDescription {
-            cv: info.cv.to_bytes(),
-            anchor: info.anchor.to_bytes(),
-            nullifier: info.nullifier.0,
-            rk: info.rk.0.to_bytes(),
-            zkproof: info.zkproof,
-        }
-    }
-
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_all(&self.cv)?;
-        writer.write_all(&self.anchor)?;
-        writer.write_all(&self.nullifier)?;
-        writer.write_all(&self.rk)?;
-        writer.write_all(&self.zkproof)
-    }
-}
-
-#[derive(Clone)]
-pub struct OutputDescription {
-    pub cv: [u8; 32],
-    pub cmu: [u8; 32],
-    pub ephemeral_key: [u8; 32],
-    pub enc_ciphertext: [u8; 580],
-    pub out_ciphertext: [u8; 80],
-    pub zkproof: [u8; GROTH_PROOF_SIZE],
-}
-
-impl OutputDescription {
-    pub fn from(
-        info: &zcash_primitives::transaction::components::OutputDescription,
-    ) -> OutputDescription {
-        OutputDescription {
-            cv: info.cv.to_bytes(),
-            cmu: info.cmu.to_bytes(),
-            ephemeral_key: info.ephemeral_key.to_bytes(),
-            enc_ciphertext: info.enc_ciphertext,
-            out_ciphertext: info.out_ciphertext,
-            zkproof: info.zkproof,
-        }
-    }
-
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_all(&self.cv)?;
-        writer.write_all(&self.cmu)?;
-        writer.write_all(&self.ephemeral_key)?;
-        writer.write_all(&self.enc_ciphertext)?;
-        writer.write_all(&self.out_ciphertext)?;
-        writer.write_all(&self.zkproof)
-    }
-}
-
 impl<P: consensus::Parameters, R: RngCore + CryptoRng> Builder<P, R> {
-    /// Creates a new `Builder` targeted for inclusion in the block with the given height
+    /// Creates a new [`Builder`] targeted for inclusion in the block with the given height
     /// and randomness source, using default values for general transaction fields.
     ///
     /// # Default values
@@ -551,6 +138,8 @@ impl<P: consensus::Parameters, R: RngCore + CryptoRng> Builder<P, R> {
     ) -> Result<(), Error> {
         // Consistency check: all anchors must equal the first one
         //Fixme: add this later when we get info from chain
+        // so that the tests don't fail
+        // see: https://github.com/Zondax/ledger-zcash-rs/issues/5
         self.anchor = Some(bls12_381::Scalar::one());
         /*
         let cmu = Node::new(note.cmu().into());
@@ -623,26 +212,13 @@ impl<P: consensus::Parameters, R: RngCore + CryptoRng> Builder<P, R> {
         Ok(())
     }
 
-    /// Sets the Sapling address to which any change will be sent.
+    /// Prepares a transaction to be transmitted to the HSM from the configured spends and outputs.
     ///
-    /// By default, change is sent to the Sapling address corresponding to the first note
-    /// being spent (i.e. the first call to [`Builder::add_sapling_spend`]).
-    /*
-    pub fn send_change_to(
-        &mut self,
-        ovk: OutgoingViewingKey,
-        to: PaymentAddress,
-        rcv: jubjub::Fr,
-        rcm: jubjub::Fr,
-        esk: jubjub::Fr,
-    ) {
-        self.change_address = Some((ovk, to, rcv, rcm, esk));
-    }
-    */
-    /// Builds a transaction from the configured spends and outputs.
+    /// Upon success, returns the structure that can be serialized in in the format understood by the HSM
+    /// and subsequently transmitted via the appropriate method.
     ///
-    /// Upon success, returns a tuple containing the final transaction, and the
-    /// [`TransactionMetadata`] generated during the build process.
+    /// After having retrieved the signatures from the HSM and having applied them with the appropriate
+    /// methods of the builder, it's possible to retrieve the final signature using [`Builder::finalize`]
     ///
     /// `consensus_branch_id` must be valid for the block height that this transaction is
     /// targeting. An invalid `consensus_branch_id` will *not* result in an error from
@@ -651,7 +227,7 @@ impl<P: consensus::Parameters, R: RngCore + CryptoRng> Builder<P, R> {
     pub fn build(
         &mut self,
         consensus_branch_id: consensus::BranchId,
-        prover: &impl TxProver,
+        prover: &impl HsmTxProver,
     ) -> Result<HsmTxData, Error> {
         //
         // Consistency checks
@@ -790,6 +366,7 @@ impl<P: consensus::Parameters, R: RngCore + CryptoRng> Builder<P, R> {
         })
     }
 
+    /// Attempt to apply the signatures for the transparent components of the transaction
     pub fn add_signatures_transparant(
         &mut self,
         signatures: Vec<secp256k1::Signature>, //get from ledger
@@ -799,6 +376,7 @@ impl<P: consensus::Parameters, R: RngCore + CryptoRng> Builder<P, R> {
             .apply_signatures(signatures, &mut self.mtx, consensus_branch_id)
     }
 
+    /// Attempt to apply the signatures for the shielded components of the transaction
     pub fn add_signatures_spend(
         &mut self,
         sign: Vec<Signature>, //get from ledger
@@ -834,6 +412,10 @@ impl<P: consensus::Parameters, R: RngCore + CryptoRng> Builder<P, R> {
         }
     }
 
+    /// Finalize the transaction, after having obtained all the signatures from the the HSM.
+    ///
+    /// Upon success, returns a tuple containing the final transaction, and the [`TransactionMetadata`]
+    /// generated during the build process.
     pub fn finalize(mut self) -> Result<(Transaction, TransactionMetadata), Error> {
         let tx = self.mtx.freeze().map_err(|_| Error::Finalization)?;
         let mut tx_meta = TransactionMetadata::new();
@@ -858,6 +440,7 @@ impl<P: consensus::Parameters, R: RngCore + CryptoRng> Builder<P, R> {
     pub joinsplit_sig: Option<[u8; 64]>,
     pub binding_sig: Option<Signature>,
      */
+    /// Same as finalize, except serialized to the format understood by the JavaScript users
     pub fn finalize_js(&mut self) -> Result<Vec<u8>, Error> {
         let mut txdata_copy = TransactionData::new();
         txdata_copy.version = self.mtx.version;
